feat(nimbus): Integrate segment retrieval function from metric-config-parser
Because

* experimenter was manually parsing segments when it could just import them from metric-config-parser

This commit

* replaces manual segment parsing with metric-config-parser’s retrieval function

Fixes #11623
