+++ experimenter/experimenter/nimbus_ui_new/tests/test_forms.py
@@ -15,7 +15,7 @@ from experimenter.openidc.tests.factories import UserFactory
 from experimenter.outcomes import Outcomes
 from experimenter.outcomes.tests import mock_valid_outcomes
 from experimenter.segments import Segments
-from experimenter.segments.tests import mock_get_segments
+from experimenter.segments.tests.mock_segments import mock_get_segments
 
 
 class RequestFormTestCase(TestCase):
