+++ experimenter/experimenter/nimbus_ui_new/tests/test_views.py
@@ -24,7 +24,7 @@ from experimenter.outcomes import Outcomes
 from experimenter.outcomes.tests import mock_valid_outcomes
 from experimenter.projects.tests.factories import ProjectFactory
 from experimenter.segments import Segments
-from experimenter.segments.tests import mock_get_segments
+from experimenter.segments.tests.mock_segments import mock_get_segments
 from experimenter.targeting.constants import TargetingConstants
